use data for notes from .env manage-tasks.py instead many python files
