+++ manage-tasks.py
@@ -1,23 +1,95 @@
 import argparse
+import os
 import time
-from logging import getLogger, basicConfig, StreamHandler, INFO
+from logging import getLogger
 
-from producer import load_producer
-from tatau_core.contract import NodeContractInfo
-from tatau_core.tatau.models import TaskDeclaration, TaskAssignment, VerificationAssignment
-from tatau_core.tatau.node import Producer
-
+import numpy as np
 from termcolor import colored
 
-basicConfig(
-    format='%(message)s',
-    level=INFO,
-    handlers=[
-        StreamHandler()
-    ],
-)
+from producer import load_producer
+from tatau_core.contract import NodeContractInfo, poa_wrapper
+from tatau_core.nn.tatau.model import Model, TrainProgress
+from tatau_core.tatau.models import TaskDeclaration, TaskAssignment, VerificationAssignment
+from tatau_core.tatau.models import TrainModel, Dataset
+from tatau_core.utils.ipfs import IPFS
+from tatau_core.utils.logging import configure_logging
+
+configure_logging(__name__)
 
 logger = getLogger(__name__)
+
+
+def train_local(x_train_path, y_train_path, x_test_path, y_test_path, model_path, batch_size, epochs):
+    model = Model.load_model(path=model_path)
+
+    x_train = np.load(x_train_path)
+    y_train = np.load(y_train_path)
+    x_test = np.load(x_test_path)
+    y_test = np.load(y_test_path)
+
+    class LocalProgress(TrainProgress):
+        def progress_callback(self, progress):
+            logger.info("Progress: {:.2f}".format(progress))
+
+    history = model.train(x=x_train, y=y_train, batch_size=batch_size, nb_epochs=epochs, train_progress=LocalProgress())
+
+    print(history)
+    loss, acc = model.eval(x=x_test, y=y_test)
+    print('loss({}):{}, acc({}):{}'.format(loss.__class__.__name__, loss, acc.__class__.__name__, acc))
+
+
+def train_remote(x_train_path, y_train_path, x_test_path, y_test_path, args):
+    logger.info("Start remote train")
+
+    producer = load_producer()
+
+    logger.info("Generate initial model weights")
+    model = Model.load_model(path=args.path)
+
+    initial_weights_path = "/tmp/tatau_initial_weights"
+    model.save_weights(initial_weights_path)
+
+    ipfs = IPFS()
+    logger.info("Upload weights to IPFS")
+    initial_weights_file = ipfs.add_file(initial_weights_path)
+
+    os.unlink(initial_weights_path)
+
+    dataset_name = os.path.basename(args.dataset)
+
+    dataset = Dataset.upload_and_create(
+        db=producer.db,
+        encryption=producer.encryption,
+        name=dataset_name,
+        x_train_path=x_train_path,
+        y_train_path=y_train_path,
+        x_test_path=x_test_path,
+        y_test_path=y_test_path,
+        minibatch_size=1000
+    )
+
+    logger.info('Dataset created: {}'.format(dataset))
+    logger.info('Create model')
+    train_model = TrainModel.upload_and_create(
+        name=args.name,
+        code_path=args.path
+    )
+
+    logger.debug('Model created: {}'.format(train_model))
+
+    logger.info('Create train job')
+    task = TaskDeclaration.create(
+        producer_id=producer.asset_id,
+        dataset_id=dataset.asset_id,
+        train_model_id=train_model.asset_id,
+        workers_needed=args.workers,
+        verifiers_needed=args.verifiers,
+        batch_size=args.batch,
+        epochs=args.epochs,
+        weights=initial_weights_file.multihash
+    )
+
+    logger.debug('Train job created: {}'.format(task))
 
 
 def red(text, on_color=None, attrs=None):
@@ -77,7 +149,8 @@ def get_progress_data(task_declaration):
         data['train_result'] = task_declaration.weights
 
     for td in TaskDeclaration.get_history(task_declaration.asset_id):
-        if td.loss and td.accuracy and td.state in [TaskDeclaration.State.EPOCH_IN_PROGRESS, TaskDeclaration.State.COMPLETED]:
+        if td.loss and td.accuracy and td.state in [TaskDeclaration.State.EPOCH_IN_PROGRESS,
+                                                    TaskDeclaration.State.COMPLETED]:
             if td.state == TaskDeclaration.State.EPOCH_IN_PROGRESS:
                 epoch = td.current_epoch - 1
             else:
@@ -218,17 +291,113 @@ def monitor_task(asset_id):
             break
 
 
-def main():
-    parser = argparse.ArgumentParser(description='Monitor Task')
+def load_wallet_credentials(account_address_var_name, storage_path_var_name):
+    address = os.getenv(account_address_var_name)
+    if address is None:
+        raise ValueError('{} is not specified'.format(account_address_var_name))
 
-    parser.add_argument('-k', '--key', default='producer', metavar='KEY', help='producer RSA key name')
-    parser.add_argument('-t', '--task', metavar='TASK_ASSET', help='asset id of task declaration')
+    storage_path = os.getenv(storage_path_var_name)
+    if storage_path is None:
+        raise ValueError('{} is not specified'.format(storage_path_var_name))
+
+    dir_name = address.replace('0x', '')
+    with open(os.path.join(storage_path, dir_name, 'wallet.json'), 'r') as f:
+        wallet = f.read()
+
+    with open(os.path.join(storage_path, dir_name, 'wallet.pass'), 'r') as f:
+        wallet_password = f.read()
+
+    return wallet, wallet_password
+
+
+def main():
+    parser = argparse.ArgumentParser(description='Produce Task')
+
+    parser.add_argument('-c', '--command', required=True, metavar='KEY', help='add|stop|cancel|issue|deposit|monitor')
+    parser.add_argument('-k', '--key', default="producer", metavar='KEY', help='RSA key name')
+    parser.add_argument('-n', '--name', default='mnist_mlp', metavar='NAME', help='model name')
+    parser.add_argument('-p', '--path', default='examples/keras/mnist/mlp.py', metavar='PATH', help='model path')
+    parser.add_argument('-d', '--dataset', default='examples/keras/mnist', metavar='dataset', help='dataset dir')
+    parser.add_argument('-w', '--workers', default=1, type=int, metavar='WORKERS', help='workers count')
+    parser.add_argument('-v', '--verifiers', default=1, type=int, metavar='VERIFIERS', help='verifiers count')
+    parser.add_argument('-b', '--batch', default=128, type=int, metavar='BATCH_SIZE', help='batch size')
+    parser.add_argument('-e', '--epochs', default=3, type=int, metavar='EPOCHS', help='epochs')
+    parser.add_argument('-l', '--local', default=0, type=int, metavar='LOCAL', help='train model local')
+    parser.add_argument('-t', '--task', default=None, type=str, metavar='TASK_ID', help='task declaration asset id')
+    parser.add_argument('-eth', '--eth', default=None, type=float, metavar='ETH', help='ETH for deposit or issue')
 
     args = parser.parse_args()
+
+    if args.command == 'add':
+        x_train_path = os.path.join(args.dataset, 'x_train.npy')
+        y_train_path = os.path.join(args.dataset, 'y_train.npy')
+        x_test_path = os.path.join(args.dataset, 'x_test.npy')
+        y_test_path = os.path.join(args.dataset, 'y_test.npy')
+
+        if args.local:
+            train_local(
+                x_train_path=x_train_path, y_train_path=y_train_path, x_test_path=x_test_path, y_test_path=y_test_path,
+                model_path=args.path, batch_size=args.batch, epochs=args.epochs
+            )
+        else:
+            train_remote(
+                x_train_path=x_train_path, y_train_path=y_train_path, x_test_path=x_test_path, y_test_path=y_test_path,
+                args=args
+            )
+        return
+
     load_producer()
-    monitor_task(args.task)
+    if not args.task:
+        print('task is not specified, arg: -t')
+        return
+
+    if args.command == 'cancel':
+        td = TaskDeclaration.get(args.task)
+        td.state = TaskDeclaration.State.FAILED
+        td.save()
+        print('Canceled {}'.format(td))
+        return
+
+    if args.command == 'stop':
+        td = TaskDeclaration.get(args.task)
+        td.state = TaskDeclaration.State.COMPLETED
+        td.save()
+        print('Stopped {}'.format(td))
+        return
+
+    if args.command == 'monitor':
+        monitor_task(args.task)
+        return
+
+    if args.command == 'issue':
+        if not args.eth:
+            print('balance is not specified, arg: --eth')
+            return
+
+        encrypted_key, password = load_wallet_credentials(
+            account_address_var_name='PRODUCER_ACCOUNT_ADDRESS',
+            storage_path_var_name='KEYS_PATH'
+        )
+        NodeContractInfo.configure(encrypted_key, password)
+        task_declaration = TaskDeclaration.get(args.task)
+        poa_wrapper.issue_job(task_declaration, args.balance)
+        return
+
+    if args.command == 'deposit':
+        if not args.eth:
+            print('balance is not specified, arg: --eth')
+            return
+
+        encrypted_key, password = load_wallet_credentials(
+            account_address_var_name='PRODUCER_ACCOUNT_ADDRESS',
+            storage_path_var_name='KEYS_PATH'
+        )
+
+        NodeContractInfo.configure(encrypted_key, password)
+        task_declaration = TaskDeclaration.get(args.task)
+        poa_wrapper.deposit(task_declaration, args.balance)
+        return
 
 
 if __name__ == '__main__':
     main()
-
