+++ verifier.yml
@@ -6,7 +6,7 @@ services:
     build:
       context: .
       dockerfile: ./docker/images/verifier/Dockerfile
-    command: sh -c "python verifier-no-socket.py"
+    command: sh -c "python verifier.py"
     # Support nvidia docker 2
     runtime: nvidia
     volumes:
