--- whitelist.json.example
@@ -1,4 +1,0 @@
-{
-  "estimators": [],
-  "verifiers": []
-}
